Refactor: templatesをpagesにリネーム(#35)
