+++ src/public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=6c16485f564ddcb78753",
+    "/js/app.js": "/js/app.js?id=f979d9d5cc7d44007bea",
     "/css/app.css": "/css/app.css?id=f4d8b4dcea3280933ba3"
 }
