+++ src/resources/ts/components/pages/Memo.tsx
